Copy the basic layout only the first time

+++ src/commands/ScaffoldCommand.php
@@ -97,7 +97,9 @@ class ScaffoldCommand extends Command {
         $src = __DIR__."/../layout";
         $dest = base_path()."/resources/views/".($this->prefix != "" ? $this->prefix."/" : "");
 
-        shell_exec("cp -r $src $dest");
+        if(!file_exists($dest)){
+            shell_exec("cp -r $src $dest");
+        }
     }
 
     public function makeMigration(){
